Partial hack for bug 31835 -- article area waaaay too small on Android 3 & 4
The iframe on Android 2 expands to its full size, and we let the entire document scroll around on a fixed-expanded iframe.
But on Android 3 and 4, it defaults to a couple hundred px high and doesn't auto size to fit.

Switching things around to lay it out at a fixed size within the window would be preferable but as soon as you scroll *the iframe*, tapping in the iframe breaks -- tapping on links etc hits where things are without scroll, instead of where they are now.

For now, at load time I'm just stretching the iframe's height to match the internal document height. This works well enough until you expand some sections, at which point the document gets taller and things end up scrolling within the iframe and exploding again. Sigh.

But, at least things are visible. ;)