First steps at getting app to run on apps without Google proprietary APIs (such as Amazon Kindle Fire tablet)
* switched <uses-library> manifest entry for Google Maps to indicate that it is not required -- this lets the app install and run on devices without the proprietary APIs.
* cleaner loading of the NearMeActivity class in the NearMePlugin, so we can catch a class lookup failure and return an error instead of crashing.
* added an alert() dialog on map loading failure, to indicate that yes you tried it but no it doesn't work.

The map view still loads on my Nexus One (without the actual maps, but I think that's because it needs a registered key to "really" work).
On Kindle Fire, the app installs and runs and the map view triggers the error dialog and lets you back to where you were before.