Add an artificial delay to keep font-size adjustment and history-addition from interfering.
They should not interfere, but when both are let to run at natural time, the history usually fails.
Need to delay one or the other, but delaying font-size adjustment doesn't seem to work reliably.

May indicate a race condition in Lawnchair, or something dumb we're doing, or something dumb in the webkit sql. ;)