let jenkins override version name
Nightly build show the same version number every night which is not very
helpful when testing a build. Jenkins publish environnement variables we
can use to override that version to something like:

 jenkins-2012-01-04_23-14-42_#32

Jenkins env variables are listed in its documentation:
  https://integration.mediawiki.org/ci/env-vars.html

aapt an not override versionName (nor versionCode) if it is set in the
AndroidManifest.xml. So I had to move it to project.properties.
If BUILD_ID and BUILD_NUMBER are set the version.name property is set,
and the value in project.properties will be ignored.