Use iScroll on iOS 3 (untested) or 4 (tested 4.3 sim)
This works around the lack of overflow scrolling on iOS before version 5.

Adds chrome.doScrollHack() calls when creating/refreshing scrollable views; separated
the scrollable divs from their content to make it work better with iScroll.

iScroll is not used on Android (too slow on 2.x, not needed on 3/4) or iOS 5 (not needed).
Though could add it in later to support pinch zoom. :D