Fix/work around bug 33884: reverts commit ae9606ed5627c60425c5d55905872913c2eb1d13
Earlier commit had changed the header on the settings screen to a <header> to match the others. For some reason, on Android 2.3.x this made the selects unselectable.
Undoing it keeps things working, as long as there are few enough items not to need to scroll. :)

It still leaves the 'close' box mispositioned on tablet mode, though, that may be fixable separately.