provide dataTypes for api calls
the jQuery ajax call only does an intelligent guess if no dataType
parameter is set. In network.js previously dataType was set to text
so we should respect that as the default to avoid breaking things
elsewhere. The search requests will need to ask for json to make sure
they get the expected json back (search is broken without this)

cc @4ndreaSt4gi